Miguel modal header
The header modal was added, a new js was made to be able to use more than one button to open the modal and the modal was made responsive.
Miguel took care of the html and the styles, Andres made changes to make it fit well and he did the js.